+++ fe-auth/src/app.css
@@ -1,3 +1,51 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+:root {
+  --gradient-color-1: #073b4c;
+  --gradient-color-2: #ef476f;
+}
+
+.gradient-border {
+  @apply relative;
+  @apply inline-flex;
+  @apply items-center;
+  @apply justify-center;
+}
+.gradient-border:before {
+  content: "";
+  position: absolute;
+  top: -1px;
+  left: -1px;
+  width: 100%;
+  height: 100%;
+  background-clip: content-box, border-box;
+  background-origin: border-box;
+  border: 2px double transparent;
+  border-radius: 9999px;
+  background-image: linear-gradient(#fff, #fff),
+    radial-gradient(circle at top left, #ef476f, #073b4c);
+}
+
+.header {
+  font-size: 200px;
+  font-family: "Roboto", sans-serif;
+  font-weight: 700;
+  @apply text-white;
+}
+
+.button-text {
+  @apply z-10;
+  @apply text-lg;
+  @apply bg-clip-text;
+  @apply text-transparent;
+  @apply bg-gradient-to-r;
+  @apply from-[--gradient-color-2];
+  @apply to-[--gradient-color-1];
+  @apply font-semibold;
+  @apply relative;
+  @apply px-5;
+  @apply p-1;
+  @apply rounded-full;
+}
