+++ fe-auth/src/lib/button.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+  export let href: string;
+  export let text: string;
+</script>
+
+<a {href} class="gradient-border px-5 py-2">
+  <span class="button-text">{text}</span>
+</a>
