+++ fe-auth/src/routes/+layout.svelte
@@ -2,4 +2,18 @@
   import "../app.css";
 </script>
 
-<slot />
+<div class="background">
+  <div class="header-container">
+    <slot />
+  </div>
+</div>
+
+<style>
+  .background {
+    @apply w-full;
+    @apply h-full;
+    @apply absolute;
+    @apply flex flex-row justify-center items-center;
+    background: linear-gradient(45deg, #073b4c 0%, #ef476f 100%);
+  }
+</style>
